requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0

# Data processing
datasets>=2.15.0
numpy>=1.24.0
pandas>=2.0.0

# Evaluation metrics
scikit-learn>=1.3.0
scipy>=1.11.0
rouge-score>=0.1.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment tracking
wandb>=0.16.0
tensorboard>=2.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.66.0
jsonlines>=4.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
